feat: non interrupted play track
