+++ src/utils/RNTPUtils.ts
@@ -181,3 +181,19 @@ export const songlistToTracklist = async (
     })
   );
 };
+
+export const clearPlaylistUninterrupted = async () => {
+  const currentQueue = await TrackPlayer.getQueue();
+  const currentTrackIndex = await TrackPlayer.getActiveTrackIndex();
+  if (currentTrackIndex === undefined) return;
+  const removeTrackIndices = [...Array(currentQueue.length).keys()];
+  removeTrackIndices.splice(currentTrackIndex, 1);
+  await TrackPlayer.remove(removeTrackIndices);
+};
+
+export const playSongUninterrupted = async (song: NoxMedia.Song) => {
+  const currentQueue = await TrackPlayer.getQueue();
+  await TrackPlayer.add(await songlistToTracklist([song]));
+  await TrackPlayer.skip(currentQueue.length);
+  await TrackPlayer.play();
+};
